Cleanup code and adding typescript - Second part.
