--- src/components/Loading.js
@@ -1,19 +1,0 @@
-const Loading = ({ message }) => {
-  return (
-    <>
-      <p className="loading">
-        <strong>loading {message}...</strong>
-      </p>
-
-      <style jsx>
-        {`
-          .loading {
-            text-align: center; 
-          }
-        `}
-      </style>
-    </>
-  );
-};
-
-export default Loading;
