--- src/components/RepositoryList.js
@@ -1,9 +1,0 @@
-import RepositoryCard from './RepositoryCard';
-
-export default function RepositoryList({ repositories, ...etc }) {
-	return (
-		<div {...etc}>
-			{repositories.map(repo => <RepositoryCard project={repo}/>)}
-		</div>
-	)
-}
