+++ src/utils/numbers.ts
@@ -1,0 +1,3 @@
+export const formatNumber = (x: string) => {
+    return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+}
